The value zero is special for (byte)lifetime
Mention this in the server reply section, and change the comparison
text to be correct in the face of it.  While here, use RFC 2119
language for this normative specification.  This language (or some
very close to it) was proffered by Tom Keiser with objection from
Andrew Deason during their conference call with Mike Meffie and
Mark Vitale.

[+1 by: Andrew Deason]
["happy with" by: Simon Wilkinson]

Change-Id: I1d0cba0942c6a228554ab97769c61792aab78987