Add a bit more about 'clear' vs. 'integrity'
The descriptions for the two levels are very similar, and some extra
clarification could be useful, as noted in Tom Kaiser's notes
from his conference call with Andrew Deason, Mike Meffie, and Mark Vitale.
The suggested text is tweaked slightly, and this commit adds such content
in two places, which may be overkill.

[+1 by: Andrew Deason]

Change-Id: I4436bc8fd55766881fd9051f88eb27f4c5c60676