RXGK_StartParams is not made of lists
Only two of the five elements of this structure are lists that should
be ordered by preference.  Reword slightly to say that the whole
structure represents the client's preferences, and that the two list
values are the supported values, ranked best to worst.  Use RFC 2119
language because this formatting is normative.

This was mentioned in Tom Keiser's notes from his conference call with
Andrew Deason, Mike Meffie, and Mark Vitale.

[+1 by: Andrew Deason]

Change-Id: I2fbd2c6b4e3a7a50bab09ac3179f0badf2fae813