Describe the GSSNegotiate errorcode field
Rather than standardize errors, make them application-specific.
Give examples of possible policy errors for some applications.

This issue was mentioned in the conference-call notes from Tom Keiser
with Andrew Deason, Mike Meffie, and Mark Vitale, but there was not
a resolution to the question.  Proposed text is mine.
Other options include to allocate a registry in some other document
(such as wilkinson-afs3-rxgk-afs, which is not inconsistent with
the error codes being application-specific).

["fine with" by: Andrew Deason,
 noting that the rxgk-afs draft will need to specify that a registry
 be created for these error values.]

Change-Id: I64de38fc587744389b0a54c84bf099e98f821cf5