Clarify CombineTokens' least-permissiveness
We talk about "numerical fields"; list them for clarity.
Also mention again that 0 is special for lifetimes and so the
"minimum of the values" may not be quite the right operation if the
value zero is involved.
Also note that application-specific data in the token should be
combined with a least-permissive combination.

These issues were noted in Tom Keiser's notes from his conference
call with Andrew Deason, Mike Meffie, and Mark Vitale.

[Adjusted to use Andrew Deason's proposed text about being "most restrictive",
and not call out the special-case value of 0 for lifetime and bytelife.
enctype and level are left as "numerical" fields, even though this is not
the correct treatment; discussion is still ongoing.]

Change-Id: I221a4d04b7eb8ac2de208061b9ac87a30dd4b686