DO NOT MERGE
A snapshot of the pthread-bos bits, all squashed together.
Done so that subsequent commits adding rxgk support to bos/bosserver
can be pushed to gerrit without using a merge commit or adding a
fake dependency chain to one of the patch series.

This snapshot is as of 267a362efb262e79eefc684cf6cddf5241443498.

Change-Id: I0245d598c12dc09a52da5cea26be2104cb826958