bozo: protect the allProcs list with a lock
Protect the list of processes with a mutex.  Part of bproc() is split out
into into bnode_DeleteProc() which makes bproc() a bit easier to read, and
fixes a leak of the token list associated with the bnode_proc entry.
Cron bnodes ops should probably be rewritten to eliminate the need to
export this lock.

Change-Id: I32e7d45e92b4d197aacfa5a79b947566a2c50712