bozo: pthread locking for bnodes list
bnode_Hold() is used to take a reference on a bnode before unlocking the
list.  bnode_Release() is used to drop any reference and must be called
with the bnode list unlocked.  bnode_Create() and bnode_FindInstance()
now return a referenced bnode which should be dropped via bnode_Release().

There is an additional queue entry in each bnode used for iteration in
bnode_ApplyInstance() and bnode_WaitAll() so that the bnode list lock can
be dropped while processing each bnode to prevent deadlocks with bproc().

WriteBozoFile() is modified to directly iterate the bnode list to avoid
dropping the bnode list lock.  This prevents changes while writing the
configuration file.

Change-Id: I2a7537f674dfa8883f1e243d1496661f74abfe75