bozo: pthread locking for bnodes list
bnode_Hold() is used to take a reference on a bnode before unlocking
the list.  bnode_Release() is used to drop any reference and must be
called with the bnode list unlocked.  bnode_Create() and
bnode_FindInstance() now return a referenced bnode which should be
dropped via bnode_Release().

bnode_ApplyInstance() makes a copy of the 'current' bnodes and
applies the all the requested operations to this consistent list of
bnodes.

Change-Id: I2ff9520317b934cda7ea9051dea57166410b97a7