bozo: pthread locking for bnodes list
bnode_Hold() is used to take a reference on a bnode before unlocking the
list.  bnode_Release() is used to drop any reference and must be called
with the bnode list unlocked.  bnode_Create() and bnode_FindInstance()
now return a referenced bnode which should be dropped via bnode_Release().

There is an additional lock to protect additions to the head of the list;
modify bnode_ApplyInstance() and bnode_WaitAll() to take it while processing
bnodes, to prevent deadlocks with bproc().

WriteBozoFile() is modified to directly iterate the bnode list to avoid
dropping the bnode list lock.  This prevents changes while writing the
configuration file.

Change-Id: I2ff9520317b934cda7ea9051dea57166410b97a7