Build both pthreaded and LWP rxgk library
There are no object files built by this makefile that are not
controlled by libtool, so it is easy for us to support both LWP
and pthread versions of the code.  This lets us build a librxgk.a,
providing just the security class functionality of rxgk without
the GSS negotiation portions.

If we did not build the LWP librxgk.a, we would need to sprinkle
conditionals on AFS_PTHREAD_ENV everywhere throughout the tree
where rxgk functionality is used, which is tedious.  It is much
cleaner to be able to always call such routines, even if sometimes
the implementations are stubs that return failure.  As it turns
out, most of the time the functionality is available anyway.

librxgk.a is deliberately not installed; it is provided only so
that in-tree binaries will be linkable.

Change-Id: I01c379242fed25d7aefefd4710a2349bdddbe235