Tighten up error handling for get_token
While here, call the gss tokens as "send" and "recv", instead of
"in" and "out", as that's confusing.  (The directions are different
with respect to gss_init_sec_context and RXGK_GSSNegotiate.)

Jump to cleanup labels before exiting the loop and before returning.
The recv_token from RXGK_GSSNegotiate may not be handled this way, though.

Change-Id: Ie9a7c4e2b7054496c6e9f3e2151074412da5c038