Add some guidance on nonce lengths for key negotiation
The client nonce is transmitted in the clear and as such does not
really add entropy to the generated master key.  It does help ensure
that the generated key is unique, e.g., if a kerberos service ticket
is used to establish multiple GSS security contexts, though.
Arguably more importantly, the client nonce also is included in the
MIC of the startparams, and as such serves to prevent a replay attack
of the returned MIC in concert with a MITM downgrade attack on the
actual StartParams received by the server.

The server nonce is wrapped in the GSS security context and does
contribute real entropy to the resultant key; however, only as many
bits of entropy as will be used for the key could possibly be useful.

Change-Id: I711d955d9767482de366b2be192dfdcc56c1fce7