Supply bounds for variable-length arrays
Define constants and use them.

There is no limit for the RXGK_Data typedef at this time, though that
is perhaps the most important case to consider.  Such a limit should
probably be merged in with a prospective limit on the generic OpenAFS
rx_opaque type.

The authenticator limit implicitly limits the appdata and call_numbers
fields, so separate limits are not needed there.

Change-Id: I40067823738aa8ff0bbf62d1706abe8924f34e3a