Export fortuna PRNG to the kernel where needed
Some systems do not supply a useful implementation of osi_readRandom()
but we want to use real random numbers in the kernel for setting the
RX epoch and connection ID.

Build hcrypto's fortuna PRNG into the rand-kernel interface we expose,
and export RAND_seed so that we can pass in entropy from userland.

If we are going to use the fortuna implementation, we must include the
fortuna implementation in the objects linked into the kernel module; do
this only on the architectures where it is needed.

Change-Id: I614d2bd9ac52803ec3b9572cc694cd836c8427dd