Add MIC routines to rxgk_crypto
Get and verify keyed MICs, and export a routine to get the length of
the MIC for a given key (enctype) so we can separate MIC and packet
payload.

We must define our own enctype-to-checksumtype table, as the one
built in to MIT krb5 is not usable with the checksum length querying
functionality.

Change-Id: Id58506aae9c2606d947428733a89ec368299f73f