Properly reserve space for the security object's bits
Set a header and trailer size on the connection as appropriate
for the security level.  The client does this when creating a new
connection, but the server doesn't have the information available
until it is checking the client's authentication response.

The MIC is a fixed-length header and we do not reserve any space
at the end since the MIC is not computed in-place and thus no padding
is needed.

Encrypted packets will decrypt to something that contains a pseudoheader,
so use that as the header size.  The trailer size is determined by the
ciphertext expansion.

Change-Id: I0c8b0d02f1d92aca824de9400c1c8f14146c4d17