Bring sanity to UUID handling
Do not insist on a UUID being present for the authenticator (or
go to great lengths to construct one), as this is only specified
to be present when the cache manager (or similar) is acting as the
RX client.  Instead, let the client specify a UUID parameter when
constructing a security object, optionally NULL, and store it in
the client's privateData.  Use this field in the authenticator
if it is present.

Change-Id: Ic10edf876127a9f4dfcb5f5e1ca8680bfe96fd99