Add rxgk_crypto_rfc3961
A crypto provider for rxgk which uses the in-tree rfc3961 library
to backend the actual crypto operations.

In order to use this backend, we need to link the rfc3961 library.
This reveals that we have been using the wrong rx library for those
bits we need; we should use libafsrpc for this pthreaded build.
We also pull in libafsauthent temporarily, as the rxgk server code
is currently looking up keys in KeyFileExt itself, even though that
should be the job of the application and not the security class.

Beware those enctype case statements, they may need revisiting.

Change-Id: I061008840369c0b3726745a13d384762556793c4