Convert bosserver to pthreads
Well, mostly.  We're cheating because there isn't a libtoolized
procmgmt library yet, but it should be safe for now.

Don't bother retaining a LWP version; this stuff is not super
complicated and is subject to only light concurrency.

Add a global bnode mutex to protect the global state -- the various
bnode queues (though not the bnodes themselves), some compound time
structures that should get a consistent snapshot, and global flags.
Add a condition variable to each bnode for signalling.

The bnode reference counting scheme is not touched; it's current
state is fairly bogus and probably broken even with LWP semantics.

No locking is done on the contents of an individual bnode; simultaneous
callers could introduce inconsistent state (but this was already the
case).

Do not use the softsig implementation, which is known to be buggy
with pthreads.  Instead, use a self-pipe to the bproc thread, whose
sole purpose is to wait for signals and timeouts.

Move creation of the bproc thread to nearly as late as possible
(with the creation of the BozoDaemon thread), to take advantage
of time-domain locking for initialization which is done while
single-threaded.  This allows several globals to be unlocked,
as they are read-only while threaded.

Change-Id: I2865e1fb011d611dd664261ae7a8f21f7ce8d7b0