Change rxgk_getkey_func semantics
Take kvno and enctype as pointers, so that 0 can be passed in as a
wildcard and the actual values returned back.

This is needed for using the getkey function to obtain a key with
which to encrypt a token for returning to the client.

Weaken the bound checks in dummy_getkey accordingly; the CheckResponse
caller already checks for nonzero-ness.

Change-Id: Id48409381d644990418196732830e43e5f6c173f