Provide rxgk_NewNullServerSecurityObject
RXGK presents in its token-obtaining service a new requirement
for rxnull connections, namely, it requires the ability to find
a key in which to encrypt the token being returned to the client.
(rxkad does not have this requirement, as the token is obtained
by communicating solely with the KDC, the RX server is not involved.)

Since RXGK_GSSNegotiate() is intended to be performed over an rxnull
connection, provide an extended rxnull security object which contains
per-security-class data, with the rxgk-specific data being used to
store a getkey function and rock.

In the future, we will likely provide rxgk_GetServerSecurityObjects
variants that provide both the rxnull and rxgk security object using
a given getkey function and rock.

Change-Id: Iec87756c8b6a1faf4a533914270f0df372c12784