Use getkey from the secobj for token-encrypting keys
Do not hardcode a function to obtain a long-term server key with
which to encrypt the token to be returned.  Instead, use the getkey
function from the security object of the connection to obtain such
a key.

Support is provided for rxnull and rxgk connections at present;
support for other classes could be provided in the future, but would
require modifying the security-class private data to have a way to
locate the rxgk-specific key.

Change-Id: I6cfd61312e30ed287dc7136807057296301e420d