Adjust for microtime() ABI on all XBSD
On the BSDs, struct timeval is not two 32-bit integers like our
struct clock, so the ABI is quite incompatible.  Use the native type
for the function call and translate to our local type accordingly.

Because the FreeBSD kernel module build system is drastically different
on master and openafs-stable-1_6_x, there is no workaround to remove.

Change-Id: Iab853c504fb20831f43e8be9aae47292bc4f78ea