Tidy up MakefileProto.FBSD.in
We don't support FreeBSD 6.x and older now.  As such, we can convert
entirely to the "new" world order, and not have to worry about adding
new sysnames to the conditionals every time we add them.

(The sense of the conditionals should really have been reversed anyway,
so that the default case would catch new sysnames correctly.  Too late now.)

Change-Id: Iad5d8b6edb9140b38568ab57ffecf0cfdaf1cbd3