Copy all the call numbers, not none of them
Sending uninitialized memory over to the peer is unwise, and
also causes it to send us packets that we drop on the floor.

This code is kind of broken anyway (and maybe not compliant with
the latest spec?  I thought we took out the explicit maxcall, but
maybe I'm just thinking of the ncalls).