rx: Add the PrAuthName type
This is the type specified in the RPC-L in
draft-brashear-afs3-pts-extended-names, which will hold extended
authentication names.

The structure is defined to as to be consistent with the output
of rxgen on the given RPC-L, but no RPC-L is added at this time.
Protocol additions which will use this type directly should add the
RPC-L to the corresponding protocol definition (.xg) file.

Provide helper routines to populate objects of this type, and lists
of them, and other helper routines.

A single routine suffices to free the display form of both single
PrAuthName objects and lists of the same.