Build rxgk support into libafsrpc
This is done rather differently for the shared and static libafsrpc
forms.  For the shared library, we can add a dependency on the shared
$(GSSAPI_LIBS) and link in liboafs_rxgk.la directly.  librxgk_pic.la
is linked into the static library when rxgk is enabled, and omitted
otherwise.  Careful control of what functions are exposed allows
static linking to continue to work when rxgk is disabled, though a
stub is needed for the case of rxgk_GetServerInfo, so that there
is a symbol present to satisfy the export symbol list.

Consumers of libafsrpc.a need not be modified in accordance
with this change.

Change-Id: I76c0329ba842fb0d4d66534810b114a0813c90a0