Make libtool symbol lists generated files
That live in the object directory.  For rxgk, we will need to make
the symbol list depend on a configure option, and moving the symbol lists
to the build tree seems better than having to sprinkle
LT_LDLIB_shlib_missing all over the place.

Also add dependencies on the symbols file to libraries which were missing
them.

The actual symbol lists should not be changed by this commit.

Change-Id: I1aad960c9cbfd6e13ab336e3416b05ac654605e2