Allow combined tokens for AFSCombineTokens
Restricting the token0 and token1 arguments to have come from
GSSNegotiate is overly restrictive and eliminates the benefit
of the CombineTokens RPC for use with AFS.

Permit any token valid for the vlserver as an argument.

Note that the output token from AFSCombineTokens is valid for
a fileserver only, and therefore may not be passed as an input
argument to AFSCombineTokens.

Change-Id: I4fb0ec36541ba7ae5a48c15dea5d1ee6ae2a084b