pioctl.c: removed unused variable
The 'rval' variable is only actually used in the LINUX20 case;
in all other places where it is assigned a value, that value
ends up in 'errcode' (if non-zero).  This also has the
side effect of eliminating the use of the uninitialized variable
'errcode' when rval was zero.

Found by clang on FreeBSD 10.0.

Change-Id: I60c56af355fdb68752d9596ff2cd7a4259b43fe9