afs_fetchstore: use consistent cpp conditionals
The 'err' label which was recently added is only used when
AFS_64BIT_CLIENT is defined; if this code is compiled without
that symbol defined, there is a label in the code which is declared
but unused, which induces a warning from clang, turned into an
error by --enable-checking.

Found by clang on FreeBSD 10.0.