Update the server-to-server communication section
Give better overviews in more appropriate locations.
Move the motivation for some VL_RegisterAddrsAndKey into the
section about declaring rxgk support, and the API restrictions
on it into the section with the RPC-L.

Don't duplicate the statement that rxgk's security index is 4.

Specify only that confidentiality protection is required; it is
not necessary to call out rxkad specifically as unusable.

Change-Id: I8993d305c8462cfeaee75fb85ccfe27ac6f16e5d