Add AFSCONF_SECOPTS_REQUREINTEG
In some situations we want to require that there are integrity checks
applied to the connection data, even if encryption is not necessary.
For such cases where fallback to clear connections is inadvisable,
use this AFSCONF_SECOPTS_REQUREINTEG flag to require integrity protection.
Data encryption is assumed to also provide integrity protection.

Since only rxkad supports integrity protection, this translates to
rxkad_auth for now.

Change-Id: Iadd4a91112f879fd4c2c34155a1735b2bd6aebfc