bos: use integrity protection for superuser-only RPCs
When a superuser's token is used to perform an RPC over an rxkad_clear
connection, an active attacker who can inject packets along that
connection can spoof any RPC which does not check the rxkad security
level, even when the server does check that the client of the rx
connection is a superuser.

Give superusers a bit more protection and require integrity protection
on the rx connections we create to perform RPCs that may only be
performed by superusers.

Change-Id: I644bf65c20d4faaaeb2836d5582965dd2a8177e6