Remove SetCallBackKey in favor of authenticator appdata
The SetCallBackKey RPC is more complicated than necessary to
achieve its goal of binding a key for secure callbacks to the
RPCs generating those callbacks.

Instead of having an extra identity field in the token (that is
to be cached in the server's per-connection private data) and a
separate lookup table for mapping the callback identity (and cache
manager UUID) to a callback key, just provide the callback key (and
token) directly in the authenticator alongside the token.  The
authenticator is constructed anew for each connection, so a new key
can be used for each connection if desired.

Accordingly, remove the cb_id field from the token format, and
the remaining references to it.

Change-Id: I4a65756cd73174c2960086393186232fa53af487