Remove the start_time field from the token format
It essentially serves no purpose.  At best, it would help ensure
time consistency between different servers, but only in some cases,
and would be more likely to result in user frustration than corrective
action.

The start_time field in the rxgk response structure (in the core
rxgk document) is needed, as it is an input to the transport key
derivation process.  However, because that start_time is unique
per-connection and a single token can be used to create multiple
connections, there can be no correspondence between the start_time
in the response and the start_time in the token.

Change-Id: I7cdabba22c58d2d5064e93b35e03aea446d4df7f