Add afsconf_BuildDbServerSecurityObjects
Used to build the server security objects for database servers that
will accept the actual user RPCs (for the VL and PR services, etc.).
Since these security objects will be used to handle
RXGK_AFSCombineTokens requests, allow a callback to be supplied that
can be used to look up the fileserver long-term key used to encrypt
fileserver tokens.

Change-Id: I59e6bce2c1ccf8c873bda913b28f26d9c2a49f63