rxgk support for viced (and volserver)
Use afsconf_BuildFileServerSecurityObjects to get our security
objects; introspect to see if there is an rxgk object in the
output, so that we can use rxgk.

If we have rxgk capability, use an rxgk connection to the vlserver
to register our addresses.  Only printed tokens from the cell-wide
key are currently supported, and we do not support registering
per-fileserver keys.  (For one, the vldb has no way of recording
that we are rxgk-capable yet.)

Use rxgk for hpr's connection to the ptserver as well.

Use afsconf_BuildFileServerSecurityObjects to get an rxgk server
security object if possible.  This is not fully usable, though,
since we can't get a CPS for such users yet.

While here, make a file-local routine static.

The use_rxgk flag should arguably be a part of the configuration
directory state, but that can wait for later.

Change-Id: I067ce991e69cb2b2752e6808573489a0aa4f5e3c

BuildFileServerSecurityObjects for volser, too

Change-Id: I2c4446bac5b9db90d57e2993bd0ebcb379736d4b