vlserver: use an rxgk getfskey routine
In order for SRXGK_AFSCombineTokens to function properly, the rx
service providing that RPC must attach service-specific data
allowing the lookup of a token-encrypting key for a given fileserver
UUID.  This commit provides a very simple implementation of that
getfskey functionality for the vlserver.

In particular, it only supports fileservers that have the shared
cell-wide key, and only fileservers that are colocated with
database servers.  This is because it is required for a database
server to have the cell-wide key for operation, and so a fileserver
on that same host must also have access to the cell-wide key.
Since no released version of this software has database servers that
implement rxgk but fileservers that do not implement rxgk, if
rxgk is enabled, we can assume that the fileserver software supports
rxgk and has access to the shared cell-wide key.

This is the best functionality that can be supplied reliably
without changes to the vldb format (which would mark whether
individual file servers are rxgk-capable, and what key to use
for them).

Change-Id: I6f7afe34c2ec5e99e7b1b3984a92d1dda6eda2de